Refactor for prop-drilling
