+++ apps/app/src/lib/i18n/dictionaries/bn.json
@@ -1221,5 +1221,6 @@
   "View review by {name}": "{name} এর পর্যালোচনা দেখুন",
   "Results are live!": "ফলাফল প্রকাশিত হয়েছে!",
   "All participants can now view the winning proposals.": "সকল অংশগ্রহণকারী এখন বিজয়ী প্রস্তাবগুলি দেখতে পারবেন।",
-  "View public results page": "পাবলিক ফলাফল পৃষ্ঠা দেখুন"
+  "View public results page": "পাবলিক ফলাফল পৃষ্ঠা দেখুন",
+  "Untitled": "নামহীন"
 }
