+++ apps/app/src/lib/i18n/dictionaries/en.json
@@ -1236,6 +1236,7 @@
   "Budget": "Budget",
   "Category": "Category",
   "Untitled proposal": "Untitled proposal",
+  "Untitled": "Untitled",
   "Advance {title}": "Advance {title}",
   "Don't advance {title}": "Don't advance {title}",
   "+{count} More": "+{count} More",
