+++ apps/app/src/lib/i18n/dictionaries/es.json
@@ -1220,5 +1220,6 @@
   "View review by {name}": "Ver revisión de {name}",
   "Results are live!": "¡Los resultados están publicados!",
   "All participants can now view the winning proposals.": "Todos los participantes ya pueden ver las propuestas ganadoras.",
-  "View public results page": "Ver página pública de resultados"
+  "View public results page": "Ver página pública de resultados",
+  "Untitled": "Sin título"
 }
