+++ apps/app/src/lib/i18n/dictionaries/fr.json
@@ -1220,5 +1220,6 @@
   "View review by {name}": "Voir la révision de {name}",
   "Results are live!": "Les résultats sont publiés !",
   "All participants can now view the winning proposals.": "Tous les participants peuvent maintenant consulter les propositions gagnantes.",
-  "View public results page": "Voir la page publique des résultats"
+  "View public results page": "Voir la page publique des résultats",
+  "Untitled": "Sans titre"
 }
