+++ apps/app/src/lib/i18n/dictionaries/pt.json
@@ -1220,5 +1220,6 @@
   "View review by {name}": "Ver avaliação de {name}",
   "Results are live!": "Os resultados estão disponíveis!",
   "All participants can now view the winning proposals.": "Todos os participantes já podem ver as propostas vencedoras.",
-  "View public results page": "Ver página pública de resultados"
+  "View public results page": "Ver página pública de resultados",
+  "Untitled": "Sem título"
 }
