+++ apps/app/src/lib/i18n/dictionaries/so.json
@@ -1173,5 +1173,6 @@
   "PROPOSALS TO ADVANCE": "SOO JEEDIMAHA HOREY LOO QAADAYO",
   "Results are live!": "Natiijooyinka ayaa la daabacay!",
   "All participants can now view the winning proposals.": "Dhammaan ka-qaybgalayaasha hadda waxay arki karaan soo jeedimaha guuleystay.",
-  "View public results page": "Eeg bogga natiijooyinka guud"
+  "View public results page": "Eeg bogga natiijooyinka guud",
+  "Untitled": "Magac la'aan"
 }
